requirements.txt
flask==2.0.1
flask-bcrypt==1.0.1
PyJWT==2.3.0
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0